Move HasHardForkHistory to own module
