+++ ouroboros-consensus/src/Ouroboros/Consensus/Node/LedgerDerivedInfo.hs
@@ -1,7 +1,3 @@
-{-# LANGUAGE DataKinds        #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE TypeFamilies     #-}
-
 module Ouroboros.Consensus.Node.LedgerDerivedInfo (
     HasHardForkHistory(..)
   , LedgerDerivedInfo(..)
@@ -9,28 +5,7 @@ module Ouroboros.Consensus.Node.LedgerDerivedInfo (
 
 import           Ouroboros.Consensus.Block
 import           Ouroboros.Consensus.BlockchainTime
-import           Ouroboros.Consensus.HardFork.History as HardFork
-import           Ouroboros.Consensus.Ledger.Abstract
-
-class HasHardForkHistory blk where
-  -- | Type level description of the hard fork shape
-  --
-  -- The exact way in which this will used will be determined once we have the
-  -- proper hard fork combinator, but intuitively it would be something like
-  -- @'[Byron, Shelley, Goguen]@.
-  type family HardForkIndices blk :: [*]
-
-  -- | Static (ledger independent) hard fork shape
-  hardForkShape :: BlockConfig blk
-                -> HardFork.Shape (HardForkIndices blk)
-
-  -- | Ledger-dependent hard fork transitions
-  --
-  -- TODO: This should eventually obsolete 'knownSlotLengths'
-  -- TODO: This should eventually address #1205
-  hardForkTransitions :: LedgerConfig blk
-                      -> LedgerState blk
-                      -> HardFork.Transitions (HardForkIndices blk)
+import           Ouroboros.Consensus.HardFork.Abstract
 
 -- | Information the node requires to run which is dependent on the ledger state
 class HasHardForkHistory blk => LedgerDerivedInfo blk where
